Updated defs to allow the result to wrapped in envelope
